backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
boto3==1.35.90
pydantic-settings==2.7.1
python-multipart==0.0.20
mangum==0.19.0
